test_adrv9002_p: Support RX/TX NCO test
To support NCO settings, we need to add a profile which enables NCO. This
profile is bound to a test which needs to run before the NCO settings test.
Moreover, the profiles test is set to only receive the specified files in
the `lvds_test_profiles` list. This separation opens the door to add
another profile test for CMOS profiles.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>